+++ test/source/verify_compressed_cache.cpp
