+++ test/source/verify_fast_multiplication.cpp
