+++ test/source/verify_log_computation.cpp
