+++ test/source/verify_magic_division.cpp
